pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Minimal RAG with OpenAI + PostgreSQL pgvector"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "openai>=1.59.0",
  "psycopg[binary]>=3.2.0",
  "pgvector>=0.3.6",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "python-dotenv>=1.0.0",
  "pypdf>=5.1.0",
  "numpy>=2.0.0",
  "beir>=2.0.0",
  "tqdm>=4.67.0",
]

[project.scripts]
rag = "rag_system.cli:main"

[project.optional-dependencies]
benchmark-plots = ["matplotlib>=3.9.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rag_system"]
